Moves data from books -> authors table
Pulls author data out of books and puts into and Author table. Since we
already had a ton of data about Authors in our database. In order to
make use of the data we already had, we created a migration that takes
data out of the Books table and inserts it into the Authors table while
associating that book with the newly created Author.