Modifies Serializer to pluck ids only
Removes #has_many from the serializer. Using these macros goes and grabs
the entire object from the db--when in actuality we probably just want
the id, and to make a subsequent request for the next resource. This is
important because the default should be for privacy unless otherwise
required. It is especially important with User data. If a book
`belongs_to` a `User` in its serializer, there is a massive potential
for sensitive information to be leaked.